Moving shared markdown transformers to one location
