+++ app/src/app/components/note-viewer/elements/markdown/MarkdownElementComponent.tsx
@@ -14,7 +14,7 @@ import { Dialog } from '../../../../services/dialogs';
 import { NoteElement } from 'upad-parse/dist/Note';
 import { ITheme } from '../../../../types/Themes';
 import Twemoji from 'twemoji/2/twemoji.amd';
-import { evaluate as fendEvaluate } from 'fend-wasm';
+import { colourTransformer, fendTransformer } from './MarkdownTransformers';
 
 export interface IMarkdownElementComponentProps extends INoteElementComponentProps {
 	search: (query: string) => void;
@@ -339,20 +339,7 @@ export default class MarkdownElementComponent extends React.Component<IMarkdownE
 			];
 		});
 
-		extension('fend', {
-			type: 'listener',
-			listeners: {
-				'images.after': (event, text: string) =>
-					// e.g. [[1+1]] -> 2
-					text.replace(/\[\[([^]+?)\]\]/gi, (match, content) => {
-						let result = fendEvaluate(content);
-						if (result === "Interrupted") {
-							result = "<i>Calculation timed out</i>";
-						}
-						return `<span title="Generated by fend">${result}</span>`;
-					})
-			}
-		} as any);
+		extension('fend', fendTransformer);
 
 		extension('graphs', () => {
 			let matches: string[] = [];
@@ -408,15 +395,6 @@ export default class MarkdownElementComponent extends React.Component<IMarkdownE
 			];
 		});
 
-		extension('colour', {
-			type: 'listener',
-			listeners: {
-				'images.after': (event, text: string) =>
-					// e.g. c[this is green](green), or c[red](rgb(255, 0, 0))
-					text.replace(/c\[([^\]]+)]\(([^()]*(?:\([^()]*\)[^()]*)?)\)/gi, (match, content, colour) =>
-						`<span style="color: ${colour}">${content}</span>`
-					)
-			}
-		} as any);
+		extension('colour', colourTransformer);
 	}
 }
