+++ app/src/app/components/note-viewer/elements/markdown/MarkdownHelpComponent.tsx
@@ -9,7 +9,7 @@ import { IShowdownOpts } from './MarkdownElementComponent';
 import './MarkdownHelpComponent.css';
 import Async, { Props as AsyncProps } from 'react-promise';
 import { Translators } from 'upad-parse/dist';
-import { evaluate as fendEvaluate } from 'fend-wasm';
+import { colourTransformer, fendTransformer } from './MarkdownTransformers';
 
 const ModalAsync = Async as { new (props: AsyncProps<FormattingHelpNote>): Async<FormattingHelpNote> };
 
@@ -65,30 +65,9 @@ export default class MarkdownHelpComponent extends React.Component {
 			];
 		});
 
-		extension('fend', {
-			type: 'listener',
-			listeners: {
-				'images.after': (event, text: string) =>
-					// e.g. [[1+1]] -> 2
-					text.replace(/\[\[([^]+?)\]\]/gi, (match, content) => {
-						let result = fendEvaluate(content);
-						if (result === "Interrupted") {
-							result = "<i>Calculation timed out</i>";
-						}
-						return `<span title="Generated by fend">${result}</span>`;
-					})
-			}
-		} as any);
+		extension('fend', fendTransformer);
 
-		extension('colour', {
-			type: 'listener',
-			listeners: {
-				'images.after': (event, text: string) =>
-					text.replace(/c\[([^\]]+)]\(([^)]+)\)/gi, (match, content, colour) =>
-						`<span style="color: ${colour}">${content}</span>`
-					)
-			}
-		} as any);
+		extension('colour', colourTransformer);
 
 		const html = new Converter({
 			parseImgDimensions: true,
